Onboarding process for new users
Users now are added to Valour Central and are sent a message by Victor
