+++ Valour/Server/Services/UserService.cs
@@ -19,6 +19,7 @@ public class UserService
     private readonly CoreHubService _coreHub;
     private readonly NodeService _nodeService;
 
+
     /// <summary>
     /// The stored user for the current request
     /// </summary>
@@ -243,223 +244,8 @@ public class UserService
         return TaskResult.SuccessResult;
     }
 
-    public async Task<TaskResult> RegisterUserAsync(RegisterUserRequest request, HttpContext ctx)
-    {
-        if (await _db.Users.AnyAsync(x => x.Name.ToLower() == request.Username.ToLower()))
-            return new(false, "Username is taken");
-        
-        var now = DateTime.Today;
-        var age = now.Year - request.DateOfBirth.Year;
-        if (request.DateOfBirth > now.AddYears(-age)) age--;
-
-        if (age < 13)
-            return new TaskResult(false, "You must be 13 to use Valour. Sorry!");
-        
-        if (await _db.UserEmails.AnyAsync(x => x.Email.ToLower() == request.Email))
-            return new(false, "This email has already been used");
-
-        var emailValid = UserUtils.TestEmail(request.Email);
-        if (!emailValid.Success)
-            return new(false, emailValid.Message);
-
-        // Check for whole blocked emails
-        if (await _db.BlockedUserEmails.AnyAsync(x => x.Email.ToLower() == request.Email.ToLower()))
-            return new(false, "Include request in body"); // Vague on purpose
-
-        var host = request.Email.Split('@')[1];
-
-        // Check for blocked host
-        if (await _db.BlockedUserEmails.AnyAsync(x => x.Email.ToLower() == host.ToLower()))
-            return new(false, "Include request in body"); // Vague on purpose
-
-
-        var usernameValid = UserUtils.TestUsername(request.Username);
-        if (!usernameValid.Success)
-            return new(false, usernameValid.Message);
-
-        var passwordValid = UserUtils.TestPasswordComplexity(request.Password);
-        if (!passwordValid.Success)
-            return new(false, passwordValid.Message);
-
-        Referral refer = null;
-        if (request.Referrer != null && !string.IsNullOrWhiteSpace(request.Referrer))
-        {
-            request.Referrer = request.Referrer.Trim();
-            var referUser = await _db.Users.FirstOrDefaultAsync(x => x.Name.ToLower() == request.Referrer.ToLower());
-            if (referUser is null)
-                return new(false, "Referrer not found");
-
-            refer = new Referral()
-            {
-                ReferrerId = referUser.Id
-            };
-        }
-
-        byte[] salt = PasswordManager.GenerateSalt();
-        byte[] hash = PasswordManager.GetHashForPassword(request.Password, salt);
-
-        await using var tran = await _db.Database.BeginTransactionAsync();
-        
-        try
-        {
-            User user = new()
-            {
-                Id = IdManager.Generate(),
-                Name = request.Username,
-                Tag = await GetUniqueTag(request.Username),
-                TimeJoined = DateTime.UtcNow,
-                TimeLastActive = DateTime.UtcNow,
-                Compliance = true, // All new users should be compliant
-            };
-
-            _db.Users.Add(user.ToDatabase());
-            await _db.SaveChangesAsync();
-
-            if (refer != null)
-            {
-                refer.UserId = user.Id;
-                await _db.Referrals.AddAsync(refer.ToDatabase());
-            }
-
-            UserPrivateInfo userPrivateInfo = new()
-            {
-                Email = request.Email,
-                Verified = false,
-                UserId = user.Id,
-                BirthDate = request.DateOfBirth,
-                Locality = request.Locality
-            };
-
-            _db.UserEmails.Add(userPrivateInfo.ToDatabase());
-
-            Valour.Database.Credential cred = new()
-            {
-                Id = IdManager.Generate(),
-                CredentialType = Valour.Database.CredentialType.PASSWORD,
-                Identifier = request.Email,
-                Salt = salt,
-                Secret = hash,
-                UserId = user.Id
-            };
-
-            _db.Credentials.Add(cred);
-
-            var emailCode = Guid.NewGuid().ToString();
-            EmailConfirmCode confirmCode = new()
-            {
-                Code = emailCode,
-                UserId = user.Id
-            };
-
-            _db.EmailConfirmCodes.Add(confirmCode.ToDatabase());
-            await _db.SaveChangesAsync();
-
-            Response result = await SendRegistrationEmail(ctx.Request, request.Email, emailCode);
-
-            if (!result.IsSuccessStatusCode)
-            {
-                _logger.LogError($"Issue sending email to {request.Email}. Error code {result.StatusCode}.");
-                await tran.RollbackAsync();
-                return new(false, "Sorry! We had an issue emailing your confirmation. Try again?");
-            }
-        }
-        catch (Exception e)
-        {
-            await tran.RollbackAsync();
-            _logger.LogError(e.Message);
-            return new(false, "Sorry! An unexpected error occured. Try again?");
-        }
-
-        await tran.CommitAsync();
-
-        return new(true, "Success");
-    }
-
     public async Task<User> GetUserAsync(string username, string tag)
         => (await _db.Users.FirstOrDefaultAsync(x => x.Name == username.ToLower() && x.Tag == tag.ToUpper())).ToModel();
-
-    public async Task<string> GetUniqueTag(string username)
-    {
-        var existing = await _db.Users.Where(x => x.Name.ToLower() == username.ToLower()).Select(x => x.Tag).ToListAsync();
-
-        string tag;
-        
-        do
-        {
-            tag = GenerateRandomTag();
-        } while (existing.Contains(tag));
-
-        return tag;
-    }
-    
-    // TODO: Prevent the one in 1.6 million chance that you will get the tag F***, along with other 'bad words'
-    // Just passed by this and realized the chances are far higher when accounting for similar-looking characters
-    private string GenerateRandomTag()
-    {
-        return new string(Enumerable.Repeat(ISharedUser.TagChars, 4)
-            .Select(s => s[Random.Shared.Next(s.Length)]).ToArray());
-    }
-
-    public async Task<TaskResult> ResendRegistrationEmail(UserPrivateInfo userPrivateInfo, HttpContext ctx, RegisterUserRequest request)
-    {
-        await using var tran = await _db.Database.BeginTransactionAsync();
-
-        try
-        {
-            _db.EmailConfirmCodes.RemoveRange(_db.EmailConfirmCodes.Where(x => x.UserId == userPrivateInfo.UserId));
-
-            var emailCode = Guid.NewGuid().ToString();
-            EmailConfirmCode confirmCode = new()
-            {
-                Code = emailCode,
-                UserId = userPrivateInfo.UserId
-            };
-
-            _db.EmailConfirmCodes.Add(confirmCode.ToDatabase());
-            await _db.SaveChangesAsync();
-
-            Response result = await SendRegistrationEmail(ctx.Request, request.Email, emailCode);
-            if (!result.IsSuccessStatusCode)
-            {
-                _logger.LogError($"Issue sending email to {request.Email}. Error code {result.StatusCode}.");
-                await tran.RollbackAsync();
-                return new(false, "Sorry! We had an issue emailing your confirmation. Try again?");
-            }
-        }
-        catch (Exception e)
-        {
-            await tran.RollbackAsync();
-            _logger.LogError(e.Message);
-            return new(false, "Sorry! An unexpected error occured. Try again?");
-        }
-
-        await tran.CommitAsync();
-
-        return new(true, "Success");
-    }
-
-    private static async Task<Response> SendRegistrationEmail(HttpRequest request, string email, string code)
-    {
-        var host = request.Host.ToUriComponent();
-        string link = $"{request.Scheme}://{host}/api/users/verify/{code}";
-
-        string emsg = $@"<body>
-                                  <h2 style='font-family:Helvetica;'>
-                                    Welcome to Valour!
-                                  </h2>
-                                  <p style='font-family:Helvetica;>
-                                    To verify your new account, please use the following link: 
-                                  </p>
-                                  <p style='font-family:Helvetica;'>
-                                    <a href='{link}'>Verify</a>
-                                  </p>
-                                </body>";
-
-        string rawmsg = $"Welcome to Valour!\nTo verify your new account, please go to the following link:\n{link}";
-
-        var result = await EmailManager.SendEmailAsync(email, "Valour Registration", rawmsg, emsg);
-        return result;
-    }
 
     public async Task<TaskResult<User>> UpdateAsync(User updatedUser)
     {
