+++ Valour/Shared/Models/ISharedUser.cs
@@ -11,6 +11,8 @@ namespace Valour.Shared.Models;
 
 public interface ISharedUser : ISharedItem
 {
+    const long VictorUserId = 20579262493097984;
+    
     const int FLAGS_TIME_UPDATE = 0x01;
     const string TagChars = "ABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
